# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Simple Cards API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/modules/": {
            "get": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Get all user's modules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Module"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Create new module",
                "parameters": [
                    {
                        "description": "Module params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.createOrUpdateModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Module"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/modules/import/quizlet": {
            "post": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Get module with cards",
                "parameters": [
                    {
                        "description": "Import module params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.quizletImportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/modules/{module_uuid}/": {
            "get": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Get module with cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ModuleWithCards"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Update module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Module params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules.createOrUpdateModuleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Module"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "tags": [
                    "modules"
                ],
                "summary": "Delete module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/modules/{module_uuid}/cards/": {
            "get": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get all module's cards",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Add new card to module",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cards.createCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/modules/{module_uuid}/cards/{card_uuid}": {
            "put": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Update card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Card UUID",
                        "name": "card_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Card update params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cards.updateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UsersAuth": []
                    }
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Delete card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Module UUID",
                        "name": "module_uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Card UUID",
                        "name": "card_uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify user creds and login",
                "parameters": [
                    {
                        "description": "User creds",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "invalid data"
                    },
                    "401": {
                        "description": "verification failed"
                    },
                    "500": {
                        "description": "token building error"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "User creds",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.authRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "400": {
                        "description": "invalid data"
                    },
                    "409": {
                        "description": "user with same login already exists"
                    },
                    "500": {
                        "description": "token building error"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "tags": [
                    "health"
                ],
                "summary": "Check database connection",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.authRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.authResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "cards.createCardRequest": {
            "type": "object",
            "required": [
                "meaning",
                "term"
            ],
            "properties": {
                "meaning": {
                    "type": "string"
                },
                "term": {
                    "type": "string"
                }
            }
        },
        "cards.updateCardRequest": {
            "type": "object",
            "properties": {
                "meaning": {
                    "type": "string"
                },
                "term": {
                    "type": "string"
                }
            }
        },
        "entity.Card": {
            "type": "object",
            "properties": {
                "meaning": {
                    "type": "string"
                },
                "module_uuid": {
                    "type": "string"
                },
                "term": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "entity.Module": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "entity.ModuleWithCards": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Card"
                    }
                },
                "name": {
                    "type": "string"
                },
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "modules.createOrUpdateModuleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "modules.quizletImportRequest": {
            "type": "object",
            "required": [
                "module_name",
                "quizlet_module_id"
            ],
            "properties": {
                "module_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "quizlet_module_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "UsersAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}